+++ CHANGELOG.md
@@ -26,8 +26,9 @@ any parts of the framework not mentioned in the documentation should generally b
 * Don't swallow `filter[]` params when there are several
 * Fix DeprecationWarning regarding collections.abc import in Python 3.7
 * Allow OPTIONS request to be used on RelationshipView
-* Avoid raising validation error for missing fields on a PATCH
-  request for polymorphic serializers
+* Remove non-JSONAPI methods (PUT and TRACE) from ModelViewSet and RelationshipView.
+  This fix might be a **BREAKING CHANGE** if your clients are incorrectly using PUT instead of PATCH.
+* Avoid validation error for missing fields on a PATCH request using polymorphic serializers
 
 ### Deprecated